feat: add acquire.py logic, pydantic models and logging configs
Module A, the acquire module, now receives an IP, a list of IPs or a IP
network in CIDR notation and executes an nmap scan to retrieve the
characteristics of such network and return them in a JSON serializable
object.

Pydantic was implemented for modern data validation, simplifying the
parsing creation of the structures containing the scan results.

The orchestration of the project will happen in main.py. The basic
execution of module A has been implemented.

The load_config function logic has been defined. It reads from the
default config.yaml.

Logging configuration was set up through the Python JSON logger
formatter.

The pyproject.toml file was updated with the new dependencies.